### 2020-12-11 #2900 @whedon
[PRE REVIEW]: stabm: Stability Measures for Feature Selection
**Submitting author:** @bommert (<a href="http://orcid.org/0000-0002-1005-9351">Andrea Bommert</a>)
**Repository:** <a href="https://github.com/bommert/stabm" target ="_blank">https://github.com/bommert/stabm</a>
**Version:** 1.1.4
