### 2020-12-13 #2900 @marcosvital
Yes, I can do that, @Kevin-Mattheus-Moerman!

@bommert, I'll handle your submission. The first thing I am going to do is to trigger a scope query, so other editors can help me check if you submission falls into JOSS requirements.
