### 2020-12-20 #2900 @danielskatz
@marcosvital - going forward, if you can refrain from becoming the editor of papers undergoing scope discussion, it would make it easier for AEiCs to track and act on the scope discussion
