### 2020-12-27 #2900 @danielskatz
Sorry @marcosvital,  now I see that @Kevin-Mattheus-Moerman asked you to handle it this way, which I think is problematic

@Kevin-Mattheus-Moerman, I think it's the job of the AEiC to try to handle to initial scope query if needed before passing it onto an editor...
