### 2020-12-29 #2900 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman, I think it's the job of the AEiC to try to handle to initial scope query if needed before passing it onto an editor...

@danielskatz yes apologies for deviating from protocol.
