### 2021-01-04 #2900 @marcosvital
No problem! I'll wait until the scope is resolved, and can handle the submission if is considered to be on scope.
