### 2021-01-14 #2900 @marcosvital
Hi, @bommert , I will be the editor of your submission.

Please take a look at the first manuscript proof generated above, and check if everything seems to be ok with it. If you make any changes, you can use `@whedon generate pdf` to have whedon create an update version (or just let me know and I'll do it).

At the same time, we can start assigning reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this - you can paste the github profile of anyone you think that could do it (but please don't tag anyone yet).
You can also check [this list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS, and see if any of them may be suitable to review this submission.
