+++ paper/paper.bib
@@ -33,11 +33,11 @@
 @inproceedings{bommert2020adjusted,
   title={Adjusted Measures for Feature Selection Stability for Data Sets with Similar Features},
   author={Bommert, Andrea and Rahnenf{\"u}hrer, J{\"o}rg},
-  booktitle={International Conference on Machine Learning, Optimization, and Data Science - {LOD} 2020, in print},
+  booktitle={Machine Learning, Optimization, and Data Science},
   year={2020},
-  url={https://arxiv.org/abs/2009.12075}
+  pages={203--214},
+  doi={10.1007/978-3-030-64583-0_19}
 }
-
 
 @article{bommert2020benchmark,
   title={Benchmark for Filter Methods for Feature Selection in High-Dimensional Classification Data},
