+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   stabm: stabm.html
-last_built: 2020-12-08T11:26Z
+last_built: 2021-01-15T10:19Z
 
