### 2021-01-18 #2900 @bommert
Thank you for being the editor of our submission!

The manuscript looks good.

From the list of people that have agreed to review for JOSS, the following persons seem to have expertise in the field of our contribution:
road2stat, martinju, prdm0, nhejazi, kdpeterson51, mcavs, osorensen, robohyun66, clauswilke.
