### 2021-01-25 #2900 @marcosvital
Hi, @bommert, thank you for the suggestions. I'll start to contact some of the possible reviewers soon.

In the meantime, do you think you could wok on your summary in order to include a brief description of the methods for a wider audience? As stated at [JOSS documentantion](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), the paper should include:

> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
