+++ paper/paper.bib
@@ -87,7 +87,7 @@
   number={3},
   pages={297--302},
   year={1945},
-  url={https://www.jstor.org/stable/1932409}
+  doi={10.2307/1932409}
 }
 
 @article{davis2006reliable,
