### 2021-01-26 #2900 @bommert
@marcosvital We have added an explanatory example in the beginning of our summary section. Do you think that this makes the high-level description of our methods understandable better for a wider audience?
