+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   stabm: stabm.html
-last_built: 2021-01-25T17:56Z
+last_built: 2021-01-26T17:51Z
 
