+++ README.Rmd
@@ -40,6 +40,8 @@ devtools::install_github("bommert/stabm")
 ```
 
 
-## Pull requests
+## Issues
+If you encounter any problems with the software or seek support, please open an issue.
 
-Pull requests are welcome.
+## Contributions
+Pull requests are welcome. Please try to stick to the coding style used in the `stabm` package.
