### 2021-01-27 #2900 @bommert
>
>
> @bommert, as we wait for reviewers, please take a look at some of JOSS requirements for the package's documentation: https://joss.readthedocs.io/en/latest/review_checklist.html#documentation
>
> Specifically, I think you'll need to add some information at the repository related to the community guidelines:
>
> > Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

We have added the information in the readme file.
