### 2021-02-04 #3010 @whedon
[REVIEW]: stabm: Stability Measures for Feature Selection
**Submitting author:** @bommert (<a href="http://orcid.org/0000-0002-1005-9351">Andrea Bommert</a>)
**Repository:** <a href="https://github.com/bommert/stabm" target ="_blank">https://github.com/bommert/stabm</a>
**Version:** v1.2.0
