### 2021-02-05 af1e189
add phd thesis as reference
