+++ paper/paper.bib
@@ -49,6 +49,14 @@
   doi={10.1016/j.csda.2019.106839}
 }
 
+@phdthesis{bommert2020integration,
+  title={Integration of Feature Selection Stability in Model Fitting},
+  author={Andrea Bommert},
+  year={2020},
+  school={TU Dortmund University, Germany},
+  doi={10.17877/DE290R-21906}
+}
+
  @inproceedings{awada2012review,
   title = {A Review of the Stability of Feature Selection Techniques for Bioinformatics Data},
   author = {Wael Awada and Taghi M. Khoshgoftaar and David Dittman and Randall Wald and Amri Napolitano},
