+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   stabm: stabm.html
-last_built: 2021-01-27T10:22Z
+last_built: 2021-02-05T15:29Z
 
