### 2021-02-09 #3010 @osorensen
@bommert

This is related to the following point on the reviewer checklist:

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

While I think the package is well documented and has good examples, they are all very simple. It might hence be difficult for someone without too much experience in the field to actually understand the real-world analysis problems which can be solved with this package. I think it would be useful if you could add one or more slightly extended examples of analysis tasks which `stabm` can help solved. Perhaps this could be put on a separate analysis page.

Here are some ideas that came to my mind while reading through the package documentation, but feel free to choose something else:

- Use `stabm` to assess the stability of variable selection algorithms, like the lasso (`glmnet::glmnet`). You could show a small example where the lasso is fit on a number of simulated datasets coming from the same distribution, and the stability of the selected set of variables computed. See, e.g., https://doi.org/10.1002/gepi.20623 or https://doi.org/10.1111/j.1467-9868.2010.00740.x for related methods.

- Perhaps you could also use `stabm` to assess the stability of cluster solutions? For example, you could use `stats::kmeans` to partition a dataset into a fixed number of clusters, and then repeat this either for random splits of the datasets or for multiple datasets from the same distribution. Something like this is often used to check if a clustering is stable with respect to random variations in the dataset (e.g., https://elifesciences.org/articles/52677).
