### 2021-02-09 #3010 @clauswilke
I agree with @osorensen. Specifically, the one vignette provided with the package is rather short. I would suggest expanding this vignette or adding additional ones to provide a little more substance.

The other comment I have is regarding community guidelines. You have one brief sentence in the Readme, but I think it would be a good idea to expand on that a little bit, and maybe add explicit contributing and code of conduct documents.
