### 2021-02-10 615b663
add feature selection example to the vignette
