+++ DESCRIPTION
@@ -34,7 +34,9 @@ Suggests:
     ggplot2 (>= 3.0.0),
     igraph (>= 1.2.1),
     knitr,
+    mlbench,
     rmarkdown,
+    rpart,
     testthat (>= 2.0.0)
 VignetteBuilder: 
     knitr
