### 2021-02-10 #3010 @mllg
https://github.com/bommert/stabm/pull/11 adds a simple feature selection example to the vignette. I've decided to use a simple regression tree here because the `rpart` package ships with R as a recommended package.
