+++ vignettes/stabm.Rmd
@@ -65,6 +65,7 @@ Finally, `stabm` also provides a visualization of the feature sets.
 plotFeatures(feats)
 ```
 
+
 # Example: Feature Selection
 
 In this example, we will analyze the stability of the feature selection of regression trees
@@ -81,10 +82,11 @@ dataset = subset(BostonHousing2, select = -cmedv)
 We write a small function which subsamples the `BostonHousing2` data frame to `ratio` percent of the observations, fits a regression tree and then returns the used features as character vector:
 
 ```{r}
-fit_tree = function(ratio = 0.67, cp = 0.01) {
-    n = nrow(dataset)
+fit_tree = function(target = "medv", data = dataset, ratio = 0.67, cp = 0.01) {
+    n = nrow(data)
     i = sample(n, n * ratio)
-    model = rpart::rpart(medv ~ ., data = dataset, subset = i, 
+    formula = as.formula(paste(target,  "~ ."))
+    model = rpart::rpart(formula = formula, data = data, subset = i, 
       control = rpart.control(maxsurrogate = 0, cp = cp))
     names(model$variable.importance)
 }
@@ -125,6 +127,103 @@ set.seed(1)
 selected_features2 = replicate(30, fit_tree(cp = 0.02), simplify = FALSE)
 stabilityJaccard(selected_features2)
 plotFeatures(selected_features2)
+```
+
+Now, we consider a different regression problem, for which there are highly correlated features.
+Again, we repeatedly select features using regression trees:
+```{r}
+dataset2 = subset(BostonHousing2, select = -town)
+dataset2$chas = as.numeric(dataset2$chas)
+
+set.seed(1)
+selected_features3 = replicate(30, fit_tree(target = "rm", data = dataset2, cp = 0.075), 
+  simplify = FALSE)
+```
+
+We choose to assess the similarities between the features with absolute Pearson correlations, but other similarity measures could be used as well.
+The similarity values of the selected features show that the two features *medv* and *cmedv* are almost perfectly correlated:
+```{r}
+# similarity matrix
+sim.mat = abs(cor(subset(dataset2, select = -rm)))
+
+sel.feats = unique(unlist(selected_features3))
+sim.mat[sel.feats, sel.feats]
+```
+Also, each of the 30 feature sets includes either *medv* or *cmedv*:
+```{r}
+plotFeatures(selected_features3, sim.mat = sim.mat)
+```
+
+When evaluating the feature selection stability, we want that the choice of *medv* instead of *cmedv* or vice versa is not seen as a lack of stability, because they contain almost the same information.
+Therefore, we use one of the *adjusted* stability measures, see `listStabilityMeasures()` in Section *Usage*.
+```{r}
+stabilityIntersectionCount(selected_features3, sim.mat = sim.mat, N = 1000)
+```
+
+The effect of the feature similarities for stability assessment can be quantified by considering the identity matrix as similarity matrix and thereby neglecting all similarities.
+Without taking into account the feature similarities, the stability value is much lower:
+```{r}
+no.sim.mat = diag(nrow(sim.mat))
+colnames(no.sim.mat) = row.names(no.sim.mat) = colnames(sim.mat)
+stabilityIntersectionCount(selected_features3, sim.mat = no.sim.mat)
+```
+
+
+# Example: Clustering
+
+As a second example, we analyze the stability of the clusters resulting from k-means clustering.
+```{r}
+set.seed(1)
+
+# select a subset of instances for visualization purposes
+inds = sample(nrow(dataset2), 50)
[…]
