### 2021-02-15 9bad1c8
update readme: CoC and contribution guidelines
