### 2021-02-15 #3010 @mllg
> The other comment I have is regarding community guidelines. You have one brief sentence in the Readme, but I think it would be a good idea to expand on that a little bit, and maybe add explicit contributing and code of conduct documents.

Both done in https://github.com/bommert/stabm/commit/9bad1c8a2a72884fe34ccd21fb90baa69053aa74.
