### 2021-02-15 #3010 @bommert
We have extended the vignette with a feature selection and a clustering example. The extended vignette has been merged into the master.
