### 2021-02-20 #3010 @mllg
I've added a reference to the checkmate package in the paper, hope that's okay for everyone.
