### 2021-02-20 #3010 @mllg
I propose that we do a CRAN release and update the reviewed version accordingly before proceeding.
