+++ paper/paper.bib
@@ -256,3 +256,14 @@
   year={2020},
   doi={10.1016/j.softx.2020.100526}
 }
+
+@article{lang2017checkmate,
+  title = {{checkmate}: Fast Argument Checks for Defensive R Programming},
+  author = {Michel Lang},
+  journal = {The R Journal},
+  year = {2017},
+  doi = {10.32614/RJ-2017-028},
+  pages = {437--445},
+  volume = {9},
+  number = {1},
+}
