+++ paper/paper.bib
@@ -258,7 +258,7 @@
 }
 
 @article{lang2017checkmate,
-  title = {{checkmate}: Fast Argument Checks for Defensive R Programming},
+  title = {{checkmate}: Fast Argument Checks for Defensive {R} Programming},
   author = {Michel Lang},
   journal = {The R Journal},
   year = {2017},
