### 2021-02-20 #3010 @bommert
@mcavs: We have added a feature selection and a clustering example in the vignette (bommert/stabm@c4e4101b6fb66a8fb236aafc041a69745d951f54). Do these examples meet the need for illustrative examples?
