### 2021-03-08 df7ab70
create data.frame of measures during build time
