+++ DESCRIPTION
@@ -41,4 +41,5 @@ Suggests:
 VignetteBuilder: 
     knitr
 Encoding: UTF-8
+LazyData: true
 RoxygenNote: 7.1.1
