### 2021-03-29 #3010 @marcosvital
@bommert and @mllg, the manuscript looks great, so we are almost ready to go. But first, you will need archive the last release of the package (on Zenodo, figshare, or other) if you already didn't do this. After that, just let me know the archive DOI.
