### 2021-03-30 #3010 @mllg
Thanks @marcosvital.

We have released a new version to CRAN today (v1.2.0) which includes all the improvements requested by the reviewers and uploaded the same version to Zenodo: http://doi.org/10.5281/zenodo.4648087
